+++ src/modules/Inputs/ToneGenerator.cpp
@@ -93,7 +93,7 @@ bool ToneGenerator::dontUseBuffer() const
 	return true;
 }
 
-bool ToneGenerator::seek(int, bool)
+bool ToneGenerator::seek(double, bool)
 {
 	return false;
 }
